wip: refactor image component common utils
